Since there were calls to the logger before we even knew where to store the logfile, we always created an empty undefined log file in the current working directory. Now we activate the logging on the preloader and activate it explicitly by using the system property.
